// WeddingService_AdminDashboard/src/app/modules/user/modules/super-admin/modules/vendor-dashboard/component/wallet-management/model/transaction-details.model.ts
export class TransactionDetails {
    id: number;
    walletId: number;
    referenceNo: string;
    particulars: string;
    transactionDate: Date;
    transactionType: number;
    mode: string;
    amount: number;
    walletBalance: number;
    createdBy: number;
    createdAt: Date;
    updatedBy: number;
    updatedAt: Date;

}